Desktop/pdf_chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
requests
ollama